+++ python/Ganga/Lib/MonitoringServices/MSGPeek/stomp.py
@@ -84,14 +84,14 @@ def _uuid( *args ):
     (http://aspn.activestate.com/ASPN/Cookbook/Python/Recipe/213761)
     """
     
-    t = long( time.time() * 1000 )
-    r = long( random.random() * 100000000000000000L )
+    t = time.time() * 1000
+    r = random.random() * 100000000000000000
   
     try:
         a = socket.gethostbyname( socket.gethostname() )
     except:
         # if we can't get a network address, just imagine one
-        a = random.random() * 100000000000000000L
+        a = random.random() * 100000000000000000
     data = str(t) + ' ' + str(r) + ' ' + str(a) + ' ' + str(args)
     data = md5.md5(data).hexdigest()
